requirements.txt
pandas
openpyxl
python-barcode[images]